Fixed a bug due to which the extension does not work on some fanfics.
This bug was caused by lack of check of the end of the fic text in addEmptyLines and addParagraphIndents functions.
These function tried to access a symbol by the index bigger that the text length. If the last symbols of texts matched regexes in these functions, this behavior caused a problem and the function stopped working without modifying page content at all.
I added a check so that the text is modified only if match index lesser that fic text length by 3 or more.